move App.js and App.css
