+++ package.json
@@ -12,6 +12,7 @@
     "react-router-dom": "^5.2.1",
     "react-scripts": "4.0.3",
     "redux": "^4.1.1",
+    "redux-logger": "^3.0.6",
     "web-vitals": "^1.1.2"
   },
   "scripts": {
