+++ src/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
 import { Provider } from 'react-redux';
-import App from './App';
+import App from './components/App';
 import store from './redux/configureStore';
 
 ReactDOM.render(
